World end (#503)
* test
* cl
* end_city
* added setentrance command for buildings. /building setentrance %name%
* fix setentrance
* test structure
* -error_unknown_generator_option
* add fat tower
* migrate to TS and fix errors
* remove unused file from JS dir
* fix read meta fo building shemas
* fix typo
* change blockdrawer
* fix error
* block_drawer
* block_draw
* fix block_drawer
* end
* end of work with block_drawer and add bld_air block for builders
* save build
* clear gen
* ass
* not work
* fix h
* fix generator layer clusters pos
* cluster
* test
* shift
* off end

+++ node_server/db/world/chunk.ts
@@ -377,7 +377,7 @@ export class DBWorldChunk {
                 const modifiers = JSON.parse(row.data);
                 const addr = new Vector(row);
                 for(const key in modifiers) {
-                    const index = parseFloat(key);
+                    const index = parseInt(key);
                     blocks.push({
                         pos: new Vector().fromFlatChunkIndex(index),
                         addr,
